+++ changelog.md
@@ -6,11 +6,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
 ### Changed
+
+## [1.3.1] - 2022-03-01
+
+### Changed
+- [Improve API documentation to reflect query usages](https://github.com/ballerina-platform/ballerina-standard-library/issues/2524)
+
+## [1.3.0] - 2022-01-29
+
+### Changed
+- [Fix Compiler plugin crash when variable is passed for `sql:ConnectionPool` and `mysql:Options`](https://github.com/ballerina-platform/ballerina-standard-library/issues/2536)
+
+## [1.2.1] - 2022-02-03
+
+### Changed
+- [Fix Compiler plugin crash when variable is passed for `sql:connectionPool` and `mysql:Options`](https://github.com/ballerina-platform/ballerina-standard-library/issues/2536)
+
+## [1.2.0] - 2021-12-13
+
+### Changed
+- Released module on top of Swan Lake Beta6 distribution
+
+## [1.1.0] - 2021-11-20
 
 ### Added
 - [Support noAccessToProcedureBodies options in mysql connector](https://github.com/ballerina-platform/ballerina-standard-library/issues/1545)
 - [Support Mysql connector failover and retries](https://github.com/ballerina-platform/ballerina-standard-library/issues/1950)
+- [Add Tooling support for MySLQ connector](https://github.com/ballerina-platform/ballerina-standard-library/issues/2279)
+
+### Changed
+- [Change queryRow return type to anydata](https://github.com/ballerina-platform/ballerina-standard-library/issues/2390)
 
 ## [1.0.0] - 2021-10-09